subscriber: split out field filter tests
This branch moves the field filter tests to their own test file,
separate from the tests for target filters which include directives that
match all callsites. This should fix the occasional flakiness in the
field-filtering tests when the tests are executed concurrently in a
particular order.

I've run the tests in a loop for about 4 minutes on my machine, and I
haven't seen any flakiness so far.
